+++ grizzled/text/__init__.py
@@ -1,5 +1,3 @@
-# $Id$
-
 """
 The ``grizzled.text`` package contains text-related classes and modules.
 """